[service/extensions] Move the extensions package to internal
