+++ service/internal/extensions/config.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package extensions // import "go.opentelemetry.io/collector/service/extensions"
+package extensions // import "go.opentelemetry.io/collector/service/internal/extensions"
 
 import "go.opentelemetry.io/collector/component"
